Lint project
